Responsive styling fixes
Fixed some styling issues for various breakpoints. Profile image on the home page now behaves and keeps its aspect ratio across various screen sizes, which required a complete restructuring of the HTML.
Cards on individual photo pages are now more responsive so more of the photo can be seen on screen. Meta data was resized for landscape orientation to avoid it breaking and ensure it's still legible